refactor : 공통 타입 정의를 위해 types.ts 파일 추가 및 관련 컴포넌트에서 import 경로 수정
